Tests/LibWeb/Text/expected/WebAnimations/animation-methods/persist.txt

   persist() sets animation's replaceState to persist
Animations are properly replaced when covered by another animation
persist() keeps an animation from being replaced
